local-ai-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pymupdf>=1.24.0
openai>=1.30.0
python-dotenv>=1.0.0
pydantic>=2.0.0
